Refactor PostController - Only post owner can update, delete post

+++ CwkSocial.Application/Posts/CommandHandlers/DeletePostHandler.cs
@@ -40,6 +40,19 @@ namespace CwkSocial.Application.Posts.CommandHandlers
                     return result;
                 }
 
+                if(post.UserProfileId != request.UserProfileId)
+                {
+                    result.IsError = true;
+                    var error = new Error
+                    {
+                        Code = ErrorCode.PostDeletNotPossible,
+                        Message = $"Only the owner of a post can delete it"
+                    };
+                    result.Errors.Add(error);
+
+                    return result;
+                }
+
                 _ctx.Posts.Remove(post);
                 await _ctx.SaveChangesAsync();
